Node Failure and Log Based Failure functionality

+++ src/tools/mpirun.c
@@ -6,11 +6,13 @@
 #include <arpa/inet.h>
 #include <netdb.h>
 #include <limits.h>
+#include <sys/file.h>
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <assert.h>
 #include <errno.h>
+#include <math.h>
 
 #define RET_COMPLETED 101
 #define RET_INCOMPLETE 102
@@ -19,9 +21,63 @@
 #define HOST_NAME_MAX 255
 #endif
 
+double prod_pred(int b, int N, int j, double R) {
+	double temp = 1.0;
+	for (int k = 0; k < j; k++) {
+		temp = temp * ((2 * (b - k)) + (R * (N - (2 * b) + k))) / (N - k);
+	}
+	return temp;
+}
+
+double sum_pred(int b, int N, double R) {
+	double temp = 1.0;
+	for (int j = 1; j <= b; j++) {
+		temp = temp + prod_pred(b, N, j, R);
+	}
+	return temp;
+}
+
+static char *parep_mpi_itoa(int num, char *str, int base) {
+	int isNegative = 0;
+	int i = 0;
+	if(num == 0) {
+		str[i++] = '0';
+		str[i] = '\0';
+		return str;
+	}
+	if(num < 0) {
+		isNegative = 1;
+		num = -num;
+	}
+	while(num != 0) {
+		int digit = num % 10;
+		str[i++] = '0' + digit;
+		num = num / 10;
+	}
+	if(isNegative) {
+		str[i++] = '-';
+	}
+	str[i] = '\0';
+	int start = 0;
+	int end = i-1;
+	while(start < end) {
+		char temp = str[start];
+		str[start] = str[end];
+		str[end] = temp;
+		start++;
+		end--;
+	}
+	return str;
+}
+
 extern char **environ;
 int main(int argc, char **argv) {
+	int down_procs = 0;
+	double ckpt_time,mtbf,recall,a,b;
+	recall = 0;
+	char **down_nodes;
 	int parep_mpi_size;
+	int nC, nR;
 	int parep_mpi_ntasks_per_node;
 	int parep_mpi_node_num;
 	int parep_mpi_node_size;
@@ -31,9 +87,25 @@ int main(int argc, char **argv) {
 		return 1;
 	}
 	
+	if(getenv("PAREP_MPI_PREDICT_RECALL") != NULL) {
+		double rec = atof(getenv("PAREP_MPI_PREDICT_RECALL"));
+		recall = rec;
+	}
+	
+	if(getenv("PAREP_MPI_CKPT_TIME") != NULL) {
+		double c = atof(getenv("PAREP_MPI_CKPT_TIME"));
+		ckpt_time = c;
+	}
+	
+	if(getenv("PAREP_MPI_MTBF") != NULL) {
+		double m = atof(getenv("PAREP_MPI_MTBF"));
+		mtbf = m;
+	}
+	
 	char job_ckpt_dir[256];
 	char nfilepath[256];
 	char snfile[256];
+	char cppath[256];
 	sprintf(job_ckpt_dir,"%s/checkpoint/%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
 	{
 		int ret;
@@ -43,7 +115,7 @@ int main(int argc, char **argv) {
 	}
 	sprintf(nfilepath,"%s/pbs_nodes",getenv("PAREP_MPI_BASE_WORKDIR"),job_ckpt_dir);
 	
-	FILE *source, *target;
+	FILE *source, *target, *cpfile;
 	char ch;
 	sprintf(snfile,"%s/checkpoint/pbs_nodes%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
 	source = fopen(snfile, "r");
@@ -52,8 +124,17 @@ int main(int argc, char **argv) {
 		return 1;
 	}
 	
+	sprintf(cppath,"%s/checkpoint/nodecopy%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
+	cpfile = fopen(cppath, "w+");
+	if(cpfile == NULL) {
+		fclose(source);
+		printf("Error opening cpfile file.\n");
+		return 1;
+	}
+	
 	target = fopen(nfilepath, "w");
 	if (target == NULL) {
+		fclose(cpfile);
 		fclose(source);
 		printf("Error opening target file.\n");
 		return 1;
@@ -61,10 +142,12 @@ int main(int argc, char **argv) {
 	
 	while ((ch = fgetc(source)) != EOF) {
 		fputc(ch, target);
+		fputc(ch, cpfile);
 	}
 	
 	fclose(source);
 	fclose(target);
+	fclose(cpfile);
 	
 	for(int i=1; i < argc; i++) {
 		if(strcmp(argv[i],"-np") == 0) {
@@ -77,6 +160,17 @@ int main(int argc, char **argv) {
 			return 1;
 		}
 	}
+	
+	if((getenv("CMP_RATIO") != NULL)  &&(getenv("REP_RATIO") != NULL)) {
+		int cr = atoi(getenv("CMP_RATIO"));
+		int rr = atoi(getenv("REP_RATIO"));
+		nR = (parep_mpi_size*rr)/(rr+cr);
+	} else if(getenv("COMP_PER_REP") != NULL) {
+		int cpr = atoi(getenv("COMP_PER_REP"));
+		if(cpr > 0) nR = parep_mpi_size/(1+cpr);
+		else nR = 0;
+	}
+	nC = parep_mpi_size - nR;
 	
 	parep_mpi_node_id = atoi(getenv("SLURM_NODEID"));
 	parep_mpi_ntasks_per_node = atoi(getenv("SLURM_NTASKS_PER_NODE"));
@@ -167,9 +261,128 @@ int main(int argc, char **argv) {
 			newargv[argc+1] = NULL;
 			execve(exec,newargv,environ);
 		}
+	} else {
+		failure_injector_pid = (pid_t)-1;
 	}
 	
 	do {
+		for(int i=1; i < argc; i++) {
+			if(strcmp(argv[i],"-np") == 0) {
+				FILE *downfile;
+				char downpath[256];
+				int nprocs;
+				char buf[20];
+				do {
+					down_nodes = NULL;
+					down_procs = 0;
+					sprintf(downpath,"%s/checkpoint/downfile%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
+					downfile = fopen(downpath, "r");
+					if((downfile != NULL) || ((downfile == NULL) && (errno == ENOENT))) {
+						char line[4096];
+						down_nodes = (char **)malloc(sizeof(char *)*parep_mpi_size);
+						for(int k=0; k < parep_mpi_size; k++) down_nodes[k] = NULL;
+						if(downfile != NULL) {
+							int lockret;
+							do {
+								lockret = flock(fileno(downfile), LOCK_EX);
+							} while(lockret != 0);
+							int index = 0;
+							while (fgets(line, sizeof(line), downfile)) {
+								down_nodes[index] = strdup(line);
+								if((down_nodes[index][0] != '\0') && (down_nodes[index][0] != '\n')) down_procs += parep_mpi_ntasks_per_node;
+								index++;
+							}
+							do {
+								lockret = flock(fileno(downfile), LOCK_UN);
+							} while(lockret != 0);
+							fclose(downfile);
+						}
+					}
+					nprocs = parep_mpi_size - down_procs;
+					parep_mpi_itoa(nprocs, buf, 10);
+					strcpy(argv[i+1],buf);
+					nR = parep_mpi_size - nC - down_procs;
+					sprintf(temp,"%d",parep_mpi_node_num - (down_procs/parep_mpi_ntasks_per_node));
+					setenv("PAREP_MPI_NODE_NUM",temp,1);
+					sprintf(temp,"%d",parep_mpi_size - down_procs);
+					setenv("PAREP_MPI_SIZE",temp,1);
+					if(nR < 0) {
+						if(failure_injector_pid != -1) {
+							int ret = kill(failure_injector_pid,SIGUSR2);
+						}
+						sleep(30);
+					} else {
+						unsetenv("COMP_PER_REP");
+						unsetenv("CMP_RATIO");
+						unsetenv("REP_RATIO");
+						if (nR > 0) {
+							parep_mpi_itoa(nC, buf, 10);
+							setenv("CMP_RATIO", buf, 1);
+							parep_mpi_itoa(nR, buf, 10);
+							setenv("REP_RATIO", buf, 1);
+							if(getenv("PAREP_MPI_COMPUTE_CKPT") != NULL) {
+								if(!strcmp(getenv("PAREP_MPI_COMPUTE_CKPT"),"1")) {
+									b = nR;
+									a = nC + nR - (2*b);
+									double Eab = sum_pred(b,nC+nR,recall);
+									double Mintr = mtbf*Eab;
+									double Ckpt = ckpt_time*((double)(a+b))/((double)(nC+nR));
+									double Topt = sqrt(2*Mintr*Ckpt);
+									parep_mpi_itoa(((int)(floor(Topt))), buf, 10);
+									setenv("PAREP_MPI_CKPT_INTERVAL", buf, 1);
+									printf("%d: Computed ckpt interval %f\n",getpid(),Topt);
+									fflush(stdout);
+								}
+							}
+						} else {
+							sprintf(buf,"%d",-1);
+							setenv("COMP_PER_REP", buf, 1);
+							if(getenv("PAREP_MPI_COMPUTE_CKPT") != NULL) {
+								if(!strcmp(getenv("PAREP_MPI_COMPUTE_CKPT"),"1")) {
+									b = nR;
+									a = nC + nR - (2*b);
+									double Eab = sum_pred(b,nC+nR,recall);
+									double Mintr = mtbf*Eab;
+									double Ckpt = ckpt_time*((double)(a+b))/((double)(nC+nR));
+									double Topt = sqrt(2*Mintr*Ckpt);
+									parep_mpi_itoa(((int)(floor(Topt))), buf, 10);
+									setenv("PAREP_MPI_CKPT_INTERVAL", buf, 1);
+									printf("%d: Computed ckpt interval %f\n",getpid(),Topt);
+									fflush(stdout);
+								}
+							}
+						}
+					}
+				} while (nR < 0);
+					
+				if(down_procs > 0) {
+					FILE *source, *target;
+					sprintf(snfile,"%s/checkpoint/pbs_nodes%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
+					target = fopen(snfile, "w");
+					source = fopen(cppath, "r");
+					
+					char line[4096];
+					
+					while (fgets(line, sizeof(line), source)) {
+						int isdown = 0;
+						for(int k = 0; k < parep_mpi_size; k++) {
+							if(down_nodes[k] != NULL) {
+								if(!strcmp(down_nodes[k],line)) {
+									isdown = 1;
+									break;
+								}
+							}
+						}
+						if(!isdown) fputs(line, target);
+					}
+					
+					fclose(source);
+					fclose(target);
+				}
+				if(down_nodes != NULL) free(down_nodes);
+				break;
+			}
+		}
 		empi_pid = fork();
 		if(empi_pid == -1) {
 			perror("fork");
@@ -178,6 +391,7 @@ int main(int argc, char **argv) {
 		
 		if(empi_pid == 0) { //CHILD PROCESS
 			char exec[1024];
+			char preloadlib[1024];
 			if(getenv("COMP_PER_REP") != NULL) {
 				if(strcmp(getenv("COMP_PER_REP"),"-1")) {
 					setenv("MV2_UD_RETRY_COUNT","32768",1);
@@ -186,6 +400,9 @@ int main(int argc, char **argv) {
 				setenv("MV2_UD_RETRY_COUNT","32768",1);
 			}
 			//setenv("MV2_USE_UD_HYBRID","0",1);
+			sprintf(preloadlib,"%s/lib/empi_hack.so",getenv("PAREP_MPI_PATH"));
+			setenv("LD_PRELOAD",preloadlib,1);
+			setenv("PAREP_MPI_PROXY_HACKED","1",1);
 			sprintf(exec,"%s/bin/mpirun",getenv("PAREP_MPI_EMPI_PATH"));
 			char **newargv;
 			newargv = (char **)malloc(sizeof(char *)*(argc+1));
@@ -232,19 +449,63 @@ int main(int argc, char **argv) {
 			} else {
 				int stat;
 				do {
-					waitpid(empi_pid,&stat,0);
+					int waitpid_ret = waitpid(empi_pid,&stat,0);
+					if(waitpid_ret < 0) {
+						if(errno == ECHILD) {
+							break;
+						}
+					}
 				} while(!WIFEXITED(stat));
 				do {
-					waitpid(coordinator_pid,&stat,0);
+					int waitpid_ret = waitpid(coordinator_pid,&stat,0);
+					if(waitpid_ret < 0) {
+						if(errno == ECHILD) {
+							break;
+						}
+					}
 				} while(!WIFEXITED(stat));
 				completed = WEXITSTATUS(stat);
 				if(completed == RET_INCOMPLETE) {
-					int ret = kill(failure_injector_pid,SIGUSR1);
+					if(failure_injector_pid != -1) {
+						int ret = kill(failure_injector_pid,SIGUSR1);
+					}
 				}
 			}
 		}
 	} while(completed == RET_INCOMPLETE);
 	assert(completed == RET_COMPLETED);
 	if(getenv("PAREP_MPI_MTBF") != NULL) kill(failure_injector_pid,SIGKILL);
+	
+	sprintf(snfile,"%s/checkpoint/pbs_nodes%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
+	source = fopen(snfile, "w");
+	if(source == NULL) {
+		printf("Error opening source file.\n");
+		return 1;
+	}
+	
+	sprintf(cppath,"%s/checkpoint/nodecopy%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
+	cpfile = fopen(cppath, "r");
+	if(cpfile == NULL) {
+		fclose(source);
+		printf("Error opening cpfile file.\n");
+		return 1;
+	}
+	
+	while ((ch = fgetc(cpfile)) != EOF) {
+		fputc(ch, source);
+	}
+	
+	fclose(source);
+	fclose(cpfile);
+	
+	int rmret;
+	do {
+		rmret = remove(cppath);
+	} while((rmret != 0) && (errno != ENOENT));
+	char downpath[256];
+	sprintf(downpath,"%s/checkpoint/downfile%s",getenv("PAREP_MPI_BASE_WORKDIR"),getenv("SLURM_JOB_ID"));
+	do {
+		rmret = remove(downpath);
+	} while((rmret != 0) && (errno != ENOENT));
 	return 0;
 }
